Move some files to new :commons_android and :commons_file modules
