+++ app/build.gradle
@@ -75,6 +75,8 @@ staticCodeAnalysis {
 }
 
 dependencies {
+    implementation project(':commons_file')
+    implementation project(':commons_android')
     implementation project(':model')
 
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
