+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/ui/home/SoundsAdapter.java
@@ -21,8 +21,8 @@ import androidx.annotation.RawRes;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.barriosnahuel.vossosunboton.R;
+import com.github.barriosnahuel.vossosunboton.commons.android.ui.Feedback;
 import com.github.barriosnahuel.vossosunboton.model.Sound;
-import com.github.barriosnahuel.vossosunboton.util.ui.Feedback;
 
 import java.io.File;
 import java.io.FileInputStream;
